requirements.txt
aiohttp<3.6.0,>=3.3.0
async-timeout
certifi
chardet
discord.py
idna
multidict
requests
urllib3
websockets==6.0
